PDKS_Listener/requirements.txt
SQLAlchemy==1.4.41
psycopg2-binary>=2.9.9
python-dotenv==0.21.1
pyzk==0.9
python-dateutil==2.8.2 